+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Clarify why and in which situations `AsyncGroupChild::wait` may not behave as expected when cancelled.
 - Add `AsyncGroupChild::start_kill` to align with Tokio's `Child::start_kill`.
 - Change `AsyncGroupChild::kill` to also `wait()` on the child, to align with Tokio's `Child::kill`.
+- Add `ErasedChild` abstraction to allow using the same type for grouped and ungrouped children.
 
 ## v3.0.0 (2023-10-30)